Implement search on entire document
Right now is very rudimentary. First it normalizes the search
text and the body text. Creates a list with paths and booleans
indicating if there was a match.

It jumps between chapters that are True in the list. It works for
simple search terms in roman languages and english. Needs more test.