Limit to open only one file due to bug
I cannot figure out for the life of me, what is happening when
the application opens several files at the same time.

Symptoms: The Book class in one window uses 2 instances of itself.
The instance of the last file opened and the instance of the window.
So, it fails to deliver resources because obviously one of them doesn't
have them.

"What causes that?" is the question I couldn't answer.